+++ src/Traits/AccessesRawData.php
@@ -28,6 +28,19 @@ trait AccessesRawData
             if (Str::endsWith($attribute, '()')) {
                 return call_user_func([$model, str_replace('()', '', $attribute)]);
             }
+            if (Str::contains($attribute, '.')) {
+                $parts = explode('.', $attribute);
+                switch (count($parts)) {
+                    case 2:
+                        return $model->{$parts[0]}->{$parts[1]};
+                    case 3:
+                        return $model->{$parts[0]}->{$parts[1]}->{$parts[2]};
+                    case 4:
+                        return $model->{$parts[0]}->{$parts[1]}->{$parts[2]}->{$parts[2]};
+                    default:
+                        return ''; // Shrug
+                }
+            }
             return $model->{$attribute};
         }