+++ src/resources/views/role/_assignment.blade.php
@@ -1,5 +1,3 @@
-<h3>{{ __('Roles') }}</h3>
-
 @foreach($roles as $role)
     <div class="custom-control custom-switch">
         {{ Form::checkbox("roles[{$role->name}]", 1, $model->hasRole($role), ['class' => 'custom-control-input', 'id' => '__role_' . $role->id]) }}