+++ src/resources/views/user/create.blade.php
@@ -6,33 +6,26 @@
 
 @section('content')
 
-@include('appshell::user._subnav', ['active' => 'users'])
-
 {!! Form::model($user, ['route' => 'appshell.user.store', 'autocomplete' => 'off', 'class' => 'row']) !!}
 
 <div class="col-12 col-md-6 col-lg-8 col-xl-9">
-    <div class="card card-accent-success">
-        <div class="card-header">{{ __('Enter Account Details') }}</div>
+    @component(theme_widget('group'), ['accent' => 'success'])
+        @slot('title'){{ __('Enter Account Details') }}@endslot
 
-        <div class="card-body">
-            @include('appshell::user._form')
-        </div>
+        @include('appshell::user._form')
 
-        <div class="card-footer">
-            <button class="btn btn-success">{{ __('Create user') }}</button>
-            <a href="#" onclick="history.back();" class="btn btn-link text-muted">{{ __('Cancel') }}</a>
-        </div>
-    </div>
+        @slot('footer')
+                <button class="btn btn-success">{{ __('Create user') }}</button>
+                <a href="#" onclick="history.back();" class="btn btn-link text-muted">{{ __('Cancel') }}</a>
+        @endslot
+    @endcomponent
 </div>
 
 <div class="col-12 col-md-6 col-lg-4 col-xl-3">
-    <div class="card card-accent-success">
-
-        <div class="card-header">{{ __('Roles') }}</div>
-        <div class="card-body">
-            @include('appshell::role._assignment', ['model' => $user])
-        </div>
-    </div>
+    @component(theme_widget('group'), ['accent' => 'success'])
+        @slot('title'){{ __('Roles') }}@endslot
+        @include('appshell::role._assignment', ['model' => $user])
+    @endcomponent
 </div>
 
 {!! Form::close() !!}