+++ src/resources/views/user/edit.blade.php
@@ -6,8 +6,6 @@
 
 @section('content')
 
-@include('appshell::user._subnav', ['active' => 'users'])
-
 {!! Form::model($user, [
     'route' => ['appshell.user.update', $user],
     'method' => 'PUT',
@@ -17,26 +15,23 @@
 ]) !!}
 
 <div class="col-12 col-md-6 col-lg-8 col-xl-9">
-    <div class="card card-accent-secondary">
-        <div class="card-header">{{ __('User Account Data') }}</div>
+    @component(theme_widget('group'), ['accent' => 'secondary'])
+        @slot('title'){{ __('User Account Data') }}@endslot
 
-        <div class="card-body">@include('appshell::user._form')</div>
+        @include('appshell::user._form')
 
-        <div class="card-footer">
+        @slot('footer')
             <button class="btn btn-primary">{{ __('Save') }}</button>
             <a href="#" onclick="history.back();" class="btn btn-link text-muted">{{ __('Cancel') }}</a>
-        </div>
-    </div>
+        @endslot
+    @endcomponent
 </div>
 
 <div class="col-12 col-md-6 col-lg-4 col-xl-3">
-    <div class="card card-accent-success">
-
-        <div class="card-header">{{ __('Roles') }}</div>
-        <div class="card-body">
-            @include('appshell::role._assignment', ['model' => $user])
-        </div>
-    </div>
+    @component(theme_widget('group'), ['accent' => 'success'])
+        @slot('title'){{ __('Roles') }}@endslot
+        @include('appshell::role._assignment', ['model' => $user])
+    @endcomponent
 </div>
 
 {!! Form::close() !!}